+++ arpes/config.py
@@ -11,6 +11,7 @@ This module also provides functions for loading configuration
 in via external files, to allow better modularity between
 different projects.
 """
+
 from __future__ import annotations
 
 import json
@@ -359,12 +360,11 @@ def setup_logging() -> None:
         ipython = get_ipython()
     except ImportError:
         return
-    try:
-        if type(ipython) == InteractiveShell and ipython.logfile:
-            CONFIG["LOGGING_STARTED"] = True
-            CONFIG["LOGGING_FILE"] = ipython.logfile
-    except AttributeError:
-        return
+
+    if isinstance(ipython, InteractiveShell) and ipython.logfile:
+        CONFIG["LOGGING_STARTED"] = True
+        CONFIG["LOGGING_FILE"] = ipython.logfile
+
     try:
         if CONFIG["ENABLE_LOGGING"] and not CONFIG["LOGGING_STARTED"]:
             CONFIG["LOGGING_STARTED"] = True
@@ -372,7 +372,8 @@ def setup_logging() -> None:
 
             log_path = generate_logfile_path()
             log_path.parent.mkdir(exist_ok=True)
-            ipython.magic(f"logstart {log_path}")
+            if isinstance(ipython, InteractiveShell):
+                ipython.run_line_magic("logstart", log_path)
             CONFIG["LOGGING_FILE"] = log_path
     except AttributeError:
         logging.exception("Attribute Error occurs.  Check module loading for IPypthon")