+++ arpes/endstations/plugin/MBS.py
@@ -1,4 +1,5 @@
 """Implements loading the text file format for MB Scientific analyzers."""
+
 from __future__ import annotations
 
 import warnings
@@ -142,7 +143,7 @@ class MBSEndstation(HemisphericalEndstation):
         n_eV = int(attrs["no_steps"])
         idx_eV = data_array.shape.index(n_eV)
 
-        if len(data_array.ndim) == TWO_DIMENSION:
+        if data_array.ndim == TWO_DIMENSION:
             phi_axis = np.linspace(
                 float(attrs["xscalemin"]),
                 float(attrs["xscalemax"]),
@@ -150,7 +151,7 @@ class MBSEndstation(HemisphericalEndstation):
                 endpoint=False,
             )
 
-            coords = {"phi": phi_axis * np.pi / 180, "eV": eV_axis}
+            coords = {"phi": np.deg2rad(phi_axis), "eV": eV_axis}
             dims = ["eV", "phi"] if idx_eV == 0 else ["phi", "eV"]
         else:
             coords = {"eV": eV_axis}