+++ arpes/utilities/jupyter.py
@@ -7,7 +7,7 @@ import json
 import os
 import urllib.request
 from datetime import UTC
-from logging import INFO, Formatter, StreamHandler, getLogger
+from logging import DEBUG, INFO, Formatter, StreamHandler, getLogger
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -26,8 +26,8 @@ __all__ = (
     "wrap_tqdm",
 )
 
-
-LOGLEVEL = INFO
+LOGLEVELS = (DEBUG, INFO)
+LOGLEVEL = LOGLEVELS[1]
 logger = getLogger(__name__)
 fmt = "%(asctime)s %(levelname)s %(name)s :%(message)s"
 formatter = Formatter(fmt)