+++ Changes.md
@@ -20,6 +20,24 @@ Major Changes from 3.0.1
 
     - Certainly, this it is indeed a lazy and carefree approach, but it's too rough method that leads to a bugs and does not mathc the current pythonic style.
 
+  - Remove utilities/attrs.py
+
+    - The functions in this module have not been used and are unlikely to be used in the future.
+
+  - modules that use the Bokeh.
+
+    There is a dependency problem among bokeh, tornard, and Jupyter, which I cannot fix because I'm haven't use Bokeh.
+
+    - arpes/plotting/band_tool.py
+    - arpes/plotting/curvature_tool.py
+    - arpes/plotting/fit_inspection_tool.py
+    - arpes/plotting/comparison_tool.py
+    - arpes/plotting/dyn_tool.py
+    - arpes/plotting/interactive_utils.py
+    - arpes/plotting/interactive.py
+    - arpes/plotting/path_tool.py
+    - arpes/plotting/mask_tool.py
+
   - Remove overlapped_stack_dispersion_plot
     - use stack_dispersion_plot with appropriate args
   - Remove G.extent in xarray_extensions, which is not so usuful